:recycle: Refactor tag detail page into component.
