+++ src/views/Tag/TagDetailView.vue
@@ -1,53 +1,23 @@
 <template>
   <layout :callback="callback">
-    <div class="grid">
-      <div class="full-left-grid border">
-        <p><md-icon class="md-size-2x">tag</md-icon></p>
-        <p class="md-display-3">{{ tag.name }}</p>
-        <tag-box-chip :color="tag.color" :text="tag.name"></tag-box-chip>
-      </div>
-      <div class="upper-right-grid">
-        <detail-card :title="detailCardTexts.author" v-if="tag.author">
-          <UserProfileIcon
+    <tag-detail :tag="tag">
+      <template v-slot:authorPicture>
+        <UserProfileIcon
             :uid="tag.author"
             :garnBarnApiCaller="garnBarnAPICaller"
           ></UserProfileIcon>
-        </detail-card>
-        <detail-card :title="detailCardTexts.subscriber">
-          <div
-            v-if="tag.subscriber && tag.subscriber.length !== 0"
-            class="flex"
-          >
-            <UserProfileIcon
+      </template>
+      <template v-slot:subscriberPicture>
+        <UserProfileIcon
               v-for="[index, subscriberUid] of tag.subscriber.entries()"
               :key="index"
               :uid="subscriberUid"
               :garnBarnApiCaller="garnBarnAPICaller"
               class="padding"
-            >
-            </UserProfileIcon>
-          </div>
-          <div v-else>
-            <md-icon>minimize</md-icon>
-          </div>
-        </detail-card>
-        <detail-card :title="detailCardTexts.reminderTime">
-          <div v-if="tag.reminderTime && tag.reminderTime.length !== 0">
-            <md-chip
-              v-for="time in tag.reminderTime" 
-              :key="time"
-              :style="`background-color: ${tag.color} !important; color: ${getFontColor(
-                tag.color
-              )} !important`"
-              ><md-icon>notifications</md-icon>{{ getHumanReadableTime(time) }}</md-chip
-            >
-          </div>
-          <div v-else>
-            <md-icon>minimize</md-icon>
-          </div>
-        </detail-card>
-      </div>
-      <div class="lower-right-grid">
+        >
+        </UserProfileIcon>
+      </template>
+      <template v-slot:buttons>
         <md-button class="md-primary md-raised" v-on:click="edit"
           >Edit</md-button
         >
@@ -55,8 +25,8 @@
           >Delete</md-button
         >
         <md-button class="md-secondary" @click="popBack">Back</md-button>
-      </div>
-    </div>
+      </template>
+    </tag-detail>            
     <DialogBoxComponent
       :dialogBoxId="'editTagDialogBox'"
       :isCustomDialogBox="true"
@@ -138,11 +108,6 @@ import firebase from "firebase/app";
 })
 export default class TagDetailView extends Vue {
   garnBarnAPICaller: GarnBarnApi | undefined = undefined;
-  detailCardTexts = {
-    author: "Author: ",
-    subscriber: "Subscribers: ",
-    reminderTime: "Reminder Time: ",
-  };
   editing = false;
   creationType = "tag";
   firebaseUser: firebase.User | undefined = undefined;
@@ -298,46 +263,6 @@ export default class TagDetailView extends Vue {
     }
 
     return tagApi;
-  }
-
-  getHumanReadableTime(unixTime: number): string {
-    var message = "";
-
-    var day = Math.floor(unixTime / 86400);
-    unixTime -= day * 86400;
-    if (day >= 1) {
-      message += day > 1 ? `${day} Days` : `${day} Day`;
-    }
-
-    var hour = Math.floor(unixTime / 3600) % 24;
-    unixTime -= hour * 3600;
-    if (hour >= 1) {
-      message += hour > 1 ? `${hour} hours` : `${hour} hour`;
-    }
-
-    var minute = Math.floor(unixTime / 60) % 60;
-    unixTime -= minute * 60;
-    if (minute >= 1) {
-      message += minute > 1 ? `${minute} minutes` : `${minute} minute`;
-    }
-
-    return message;
-  }
-
-  hexToRgb(hex: string): Array<number> {
-    const r = parseInt(hex.slice(1, 3), 16);
-    const g = parseInt(hex.slice(3, 5), 16);
-    const b = parseInt(hex.slice(5, 7), 16);
-    return [r, g, b];
-  }
-
-  getFontColor(hex: string): string {
-    const rgb = this.hexToRgb(hex);
-    if (rgb[0] * 0.299 + rgb[1] * 0.587 + rgb[2] * 0.114 > 186) {
-      return "#000000";
-    } else {
-      return "#ffffff";
-    }
   }
 
   popBack() {
